#22 package refactor
