+++ app/src/main/java/com/example/dodo/presentation/todoadd/TodoAddViewModel.kt
@@ -1,10 +1,9 @@
 package com.example.dodo.presentation.todoadd
 
 import com.example.dodo.presentation.base.BaseViewModel
-import com.example.dodo.presentation.base.BottomSheetScreen
+import com.example.dodo.presentation.common.BottomSheetScreen
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
 import org.orbitmvi.orbit.syntax.simple.reduce
 import org.orbitmvi.orbit.viewmodel.container
 import javax.inject.Inject
